Splitted some components and optimized initial modal code
